feat: 별점 평점 0.5단위 소수점 입력 지원 및 슬라이더 바 조절기 탑재, 반개 별(🌗) 렌더링 헬퍼 결합 완료

+++ backend/src/reviews/entities/review.entity.ts
@@ -26,7 +26,7 @@ export class Review {
   @Column({ name: 'booking_id' })
   bookingId: string;
 
-  @Column({ type: 'integer' })
+  @Column({ type: 'float' })
   rating: number;
 
   @Column({ type: 'text' })
